+++ lib/core/services/User.service.dart
@@ -4,6 +4,7 @@ import 'package:kali/core/domains/userNutriscore.repository.dart';
 import 'package:kali/core/models/EditUser.formdata.dart';
 import 'package:kali/core/models/NutriScore.model.dart';
 import 'package:kali/core/models/User.model.dart';
+import 'package:kali/core/services/Authentication.service.dart';
 import 'package:kali/core/states/user.state.dart';
 
 final userService = UserService();
@@ -13,8 +14,10 @@ class UserService {
   final _userNutriscoreRepository = UserNutriscoreRepository();
 
   Future<void> refreshUser() async {
-    User? user = await _userRepository.refreshUser();
-    userState.user.value = user;
+    if (authenticationService.isAuthentified) {
+      User? user = await _userRepository.refreshUser();
+      userState.user.value = user;
+    }
   }
 
   Future<bool> canCompute() async {